requirements.txt
streamlit
pandas
numpy
openvolley-pydatavolley
scikit-learn
plotly
pyarrow
google-api-python-client
google-auth
